Home banners: same text styling + faithful preview as the store hero
Bring the store banner's treatment to the app home banners (which also
overlay text on an uploaded image):

- shared_flutter PromoBanner gains textPosition / textAlign /
  headlineScale (defaults preserve the current look); home_feed's
  image banner honours them (position, alignment, scaled headline).
- Portal /banners: text position/align/size controls appear when an
  image is uploaded, saved onto the banner.
- PromoPreview now renders an image home banner as a real text-over-image
  overlay (matching the app) honouring those styles, instead of the old
  image-on-top-with-text-below strip. Compose + View-modal previews pass
  the styles through.
- Fix headline clamp to return double in both card renderers.

+++ apps/customer/lib/features/store/store_home_screen.dart
@@ -585,7 +585,7 @@ class _BannerCarouselState extends State<_BannerCarousel> {
     final scrimDark = [const Color(0xB01E1204), const Color(0x201E1204), const Color(0x00000000)];
     final scrimLight = [const Color(0xC8FBF3E0), const Color(0x40FBF3E0), const Color(0x00FFFFFF)];
     final scrim = onLight ? scrimLight : scrimDark;
-    final headlineSize = (25.0 * b.headlineScale).clamp(14.0, 40.0);
+    final headlineSize = (25.0 * b.headlineScale).clamp(14.0, 40.0).toDouble();
 
     return GestureDetector(
       onTap: b.linkCategoryId.isEmpty ? null : () => context.push('/store/category/${b.linkCategoryId}'),
